Update Set-JiraIssue.ps1
Fixed bug on line 90.  In order to "unassign" an assignee the $assigneeString must be set to $null instead of empty string "".
Also added case for setting assignee to system default (usually unassigned, but not always as this can be updated in JIRA settings by admin users).  Default assignee value is -1.